Remove stream-internals crate
