+++ crates/msgs/Cargo.toml
@@ -16,7 +16,6 @@ rand.workspace = true
 rmp-serde.workspace = true
 schemars.workspace = true
 serde.workspace = true
-stream-internals.workspace = true
 tracing.workspace = true
 uuid.workspace = true
 validator.workspace = true
