--- crates/stream-internals/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "stream-internals"
-edition = "2024"
-license.workspace = true
-version.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-fjall-utils.workspace = true
-schemars.workspace = true
-serde.workspace = true
-
-[lints]
-workspace = true
