--- crates/stream-internals/src/lib.rs
@@ -1,3 +1,0 @@
-//! Shared bones for Stream, Queues, and (name tbd) FIFO
-
-pub mod entities;
